[MCJ-1489] refactor: MyPageParticipationQueryServiceTest fixture 분리
